fixed comparison warnings
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem-functionals/trunk@110 4028485c-44d9-4cde-a312-5a4635ee2db9

+++ dune/fem/solver/femassembler.hh
@@ -76,7 +76,7 @@ public:
     const DFSType& space = functional.space();
 
     // check that number of dofs in space is equal to vector size
-    assert( space.size() == vec.size() );
+    assert( space.size() == (int)vec.size() );
     ItType it = space.begin();
     for( ; it != space.end(); ++it )
     {
